+++ foxy_farmer/first_run/first_run_wizard.py
@@ -4,10 +4,13 @@ from pathlib import Path
 from sys import stdout, stdin, stderr
 from typing import List, Any, Dict
 
+from chia.cmds.cmds_util import get_wallet
 from chia.cmds.init_funcs import check_keys
 from chia.cmds.keys_funcs import query_and_add_private_key_seed
 from chia.cmds.units import units
 from chia.daemon.keychain_proxy import KeychainProxy
+from chia.types.blockchain_format.sized_bytes import bytes32
+from chia.util.bech32m import encode_puzzle_hash
 from chia.util.ints import uint64
 from chia.util.keychain import Keychain
 from prompt_toolkit.shortcuts import CompleteStyle
@@ -20,7 +23,11 @@ from foxy_farmer.util.bech32_address import is_valid_address
 from foxy_farmer.util.fee import is_valid_fee
 from foxy_farmer.pool.plot_nft_updater import PlotNftUpdater
 from foxy_farmer.pool.pool_joiner import PoolJoiner
-from foxy_farmer.wallet.pool_join import get_plot_nft_not_pooling_with_foxy
+from foxy_farmer.wallet.balance import ensure_balance
+from foxy_farmer.wallet.pool_join import get_plot_nft_not_pooling_with_foxy, update_foxy_config_plot_nfts_if_required, \
+    create_plot_nft, get_plot_nft_from_config
+from foxy_farmer.wallet.run_wallet import run_wallet
+from foxy_farmer.wallet.sync import wait_for_wallet_sync
 
 
 async def run_first_run_wizard(foxy_root: Path, config: Dict[str, Any], foxy_config: FoxyConfig):
@@ -29,7 +36,10 @@ async def run_first_run_wizard(foxy_root: Path, config: Dict[str, Any], foxy_con
 
         return
     print("New install detected, running first run wizard")
-    use_og_pooling: bool = await confirm(message="Do you have OG (SOLO) plots you want to pool with?", default=foxy_config["enable_og_pooling"]).unsafe_ask_async()
+    use_og_pooling: bool = await confirm(
+        message="Do you have OG (SOLO) plots you want to pool with?",
+        default=foxy_config["enable_og_pooling"],
+    ).unsafe_ask_async()
     foxy_config["enable_og_pooling"] = use_og_pooling
 
     backend_choices: List[Choice] = [
@@ -120,7 +130,16 @@ async def run_first_run_wizard(foxy_root: Path, config: Dict[str, Any], foxy_con
         if plot_nft_count == 0:
             print("No PlotNFTs found, skipping ..")
         else:
-            print(f"{'Updated' if has_plot_nfts else 'Found'} {plot_nft_count} PlotNFT{'s' if plot_nft_count > 1 else ''}")
+            print(
+                f"{'Updated' if has_plot_nfts else 'Found'} {plot_nft_count} PlotNFT{'s' if plot_nft_count > 1 else ''}")
+
+    if len(foxy_config["plot_nfts"]) == 0 and len(all_sks) > 0:
+        should_create_plot_nft: bool = await confirm(
+            message="No PlotNFTs found, do you want to create a new PlotNFT?",
+            default=True,
+        ).unsafe_ask_async()
+        if should_create_plot_nft:
+            await create_plot_nft_interactive(foxy_root, config, foxy_config, )
 
     if len(foxy_config["plot_nfts"]) > 0 and len(all_sks) > 0:
         plot_nfts_not_pooling_with_foxy = get_plot_nft_not_pooling_with_foxy(foxy_root)
@@ -144,13 +163,51 @@ async def run_first_run_wizard(foxy_root: Path, config: Dict[str, Any], foxy_con
             default=True,
         ).unsafe_ask_async()
         if should_print_login_links:
-            print("Login links for your PlotNFTs:")
             login_links = await generate_login_links(
                 KeychainProxy(log=getLogger("first_run_wizard"), local_keychain=keychain),
                 foxy_config["plot_nfts"],
             )
+            print("Login links for your PlotNFTs:")
             for launcher_id, login_link in login_links:
                 print()
-                print(f"Launcher Id: {launcher_id}")
+                print(f"Launcher ID: {launcher_id}")
                 print(f" Login Link: {login_link}")
             print()
+
+
+async def create_plot_nft_interactive(foxy_root: Path, config: Dict[str, Any], foxy_config: FoxyConfig):
+    fee_str: str = await text(
+        message="Which fee (in XCH) do you want to use?",
+        default="0",
+        validate=is_valid_fee,
+    ).unsafe_ask_async()
+    fee_raw: uint64 = uint64(int(Decimal(fee_str) * units["chia"]))
+    async with run_wallet(root_path=foxy_root, config=config) as wallet_rpc:
+        # Select wallet to sync
+        await get_wallet(foxy_root, wallet_rpc, fingerprint=None)
+        await wait_for_wallet_sync(wallet_rpc)
+        wallet_address = await wallet_rpc.get_next_address(1, new_address=False)
+
+        await ensure_balance(
+            wallet_client=wallet_rpc,
+            required_balance_mojo=fee_raw + 1,
+            message_while_waiting_closure=lambda remaining_xch: f"Not enough XCH available to create PlotNFT, please send at least {remaining_xch} XCH to {wallet_address}",
+        )
+
+        launcher_id = await create_plot_nft(wallet_rpc, fee=fee_raw)
+        if launcher_id is None:
+            return
+        await wait_for_wallet_sync(wallet_rpc)
+        update_foxy_config_plot_nfts_if_required(foxy_root, foxy_config)
+        plot_nft = get_plot_nft_from_config(foxy_root, launcher_id)
+        if plot_nft is None:
+            return
+        pool_contract_address = encode_puzzle_hash(
+            puzzle_hash=bytes32.from_hexstr(plot_nft['p2_singleton_puzzle_hash']),
+            prefix="xch",
+        )
+        print("!! Please make note of the Launcher ID and Pool Contract Address below !!")
+        print()
+        print(f"          Launcher ID: {launcher_id}")
+        print(f"Pool Contract Address: {pool_contract_address}")
+        print()